@@ -3,26 +3,29 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.5.0",
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.3",
-    "@testing-library/user-event": "^12.6.0",
-    "@types/jest": "^26.0.20",
-    "@types/node": "^12.19.14",
-    "@types/react": "^16.14.2",
-    "@types/react-dom": "^16.9.10",
-    "axios": "^0.21.1",
-    "axios-auth-refresh": "^3.0.0",
-    "bootstrap": "^4.5.3",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.4.3",
-    "react-dom": "^17.0.1",
-    "react-redux": "^7.2.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "redux-thunk": "^2.3.0",
-    "typescript": "^4.1.3",
-    "web-vitals": "^0.2.4"
+    "@reduxjs/toolkit": "^1.9.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.2.2",
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.8",
+    "@types/axios": "^0.14.0",
+    "@types/react-redux": "^7.1.15",
+    "@types/react-router-dom": "^5.1.7",
+    "axios": "^1.1.3",
+    "axios-auth-refresh": "^3.3.4",
+    "bootstrap": "^4.6.2",
+    "react": "^18.2.0",
+    "react-bootstrap": "^1.6.6",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.4.3",
+    "react-scripts": "5.0.1",
+    "redux-thunk": "^2.4.2",
+    "typescript": "^4.8.4",
+    "web-vitals": "^3.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -48,8 +51,5 @@
       "last 1 safari version"
     ]
   },
-  "devDependencies": {
-    "@types/react-redux": "^7.1.15",
-    "@types/react-router-dom": "^5.1.7"
-  }
+  "devDependencies": {}
 }
