@@ -1,31 +1,31 @@
 import { ReactElement, useEffect } from "react";
 import { useSelector } from "react-redux";
-import { Route, Switch, useHistory, useLocation } from "react-router-dom";
+import { Route, Routes, useNavigate, useLocation } from "react-router-dom";
 import { selectIsLoggedIn } from "./pages/login/store/login-selectors";
 import UsersView from "./pages/users/users-view";
 import LoginView from "./pages/login/login-page";
 
 const RouterComponent = (): ReactElement => {
   const location = useLocation();
-  const history = useHistory();
+  const history = useNavigate();
 
   const isLoggedIn = useSelector(selectIsLoggedIn);
 
   useEffect(() => {
     if (isLoggedIn && location.pathname === "/") {
-      history.push("/users");
+      history('/users');
     }
 
     if (!isLoggedIn && location.pathname !== "/") {
-      history.push("/");
+      history('/');
     }
   }, [location, isLoggedIn, history]);
 
   return (
-    <Switch>
-      <Route exact path="/" component={LoginView} />
-      <Route exact path="/users" component={UsersView} />
-    </Switch>
+    <Routes>
+      <Route path="/" element={<LoginView />} />
+      <Route path="/users" element={<UsersView />} />
+    </Routes>
   );
 };
 
